Fix file pair lists for used directories.
The key for used directories shall not be specific for the inheritance characteristics. This leads to fragmented file pair lists. Instead there should only be a single entry for each directory pair which in turn has all file pairs.

In contrast to the original inheritance information for directory pairs, more specific information needs to be curated. This is because with potentially deeper directory trees, more information is needed in order to correctly draw the deep directory dependency graphs.